@@ -1,12 +1,15 @@
 import {ChangeDetectionStrategy, Component, Input} from '@angular/core';
 
+/**
+ * Dumb component, presents background to the grid.
+ */
 @Component({
-  selector: 'plaid-planner-grid-background',
-  templateUrl: './planner-grid-background.component.html',
-  styleUrls: ['./planner-grid-background.component.scss'],
+  selector: 'plaid-grid-background',
+  templateUrl: './grid-background.component.html',
+  styleUrls: ['./grid-background.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class PlannerGridBackgroundComponent {
+export class GridBackgroundComponent {
   readonly hoursLabels: string[] = Array.from<string>({ length: 24 }).map((_, i) =>
     new Date(1970, 0, 1, i).toLocaleTimeString(undefined, { hour: 'numeric' })
   );
