@@ -3,21 +3,23 @@ import {PlaidFacade} from './plaid.facade';
 import {Worklog} from './models/worklog';
 import {DateRange} from './models/date-range';
 import {User} from './models/user';
+import {ConnectionIssueModalVisible} from './components/connection-issue-resolver/connection-issue-modal-visible';
 
+/**
+ * Application container.
+ */
 @Component({
   selector: 'plaid-root',
   templateUrl: './plaid.component.html',
   styleUrls: ['./plaid.component.scss']
 })
-
 export class PlaidComponent implements OnInit {
-  pixelsPerMinute = 2;
+  pixelsPerMinute: number;
   worklogs: Worklog[];
   loading: boolean;
   _selectedDateRange: DateRange;
   _currentUser: User;
-  changeCredentials = new EventEmitter<void>();
-  connectionIssueResolverModalVisible = false;
+  connectionIssueModalVisible = false;
 
   constructor(private facade: PlaidFacade) {}
 
@@ -26,20 +28,28 @@ export class PlaidComponent implements OnInit {
     this.facade.getAuthenticatedUser$().subscribe(user => this.currentUser = user);
     this.facade.getWorklogs$().subscribe(worklogs => this.worklogs = worklogs);
     this.facade.getWorklogsFetching$().subscribe(loading => this.loading = loading);
+    this.facade.getConnectionIssueModalVisible$()
+      .subscribe(val => this.connectionIssueModalVisible = val !== ConnectionIssueModalVisible.NONE);
   }
 
+  /**
+   * Date range visible on the grid. After range change work log entries are fetched for given range.
+   */
   set selectedDateRange(dateRange: DateRange) {
     this._selectedDateRange = dateRange;
     if (this.currentUser) {
       this.facade.fetchWorklogs(dateRange, this.currentUser);
     } else {
-      this.facade.fetchAuthenticatedUser();
+      this.facade.reconnect();
     }
   }
   get selectedDateRange(): DateRange {
     return this._selectedDateRange;
   }
 
+  /**
+   * Currently authenticated user. After change work log entries for current date range are updated.
+   */
   set currentUser(user: User) {
     this._currentUser = user;
     if (user) {
@@ -60,8 +70,11 @@ export class PlaidComponent implements OnInit {
     }
   }
 
+  changeCredentials(): void {
+    this.facade.showLoginModal();
+  }
+
   forgetAccount(): void {
-    this.facade.setAuthInfo(null);
-    this.facade.discardAuthenticatedUser();
+    this.facade.logout();
   }
 }
