@@ -2,13 +2,13 @@ import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
 import {HTTP_INTERCEPTORS, HttpClientModule} from '@angular/common/http';
 import {PlaidComponent} from './plaid.component';
-import {PlannerComponent} from './components/planner/planner.component';
+import {GridComponent} from './components/grid/grid.component';
 import {WorklogPanelComponent} from './components/worklog-panel/worklog-panel.component';
 import {AuthInterceptor} from './core/auth/auth.interceptor';
 import {ConnectionIssueResolverComponent} from './components/connection-issue-resolver/connection-issue-resolver.component';
 import {FormsModule} from '@angular/forms';
 import {DateRangePickerComponent} from './components/date-range-picker/date-range-picker.component';
-import {PlannerGridBackgroundComponent} from './components/planner-grid-background/planner-grid-background.component';
+import {GridBackgroundComponent} from './components/grid-background/grid-background.component';
 import {RefreshButtonComponent} from './components/refresh-buton/refresh-button.component';
 import {ExtHrefDirective} from './directives/ext-href.directive';
 import {NgxElectronModule} from 'ngx-electron';
@@ -19,11 +19,11 @@ import {CurrentTimeMarkerComponent} from './components/current-time-marker/curre
 @NgModule({
   declarations: [
     PlaidComponent,
-    PlannerComponent,
+    GridComponent,
     WorklogPanelComponent,
     ConnectionIssueResolverComponent,
     DateRangePickerComponent,
-    PlannerGridBackgroundComponent,
+    GridBackgroundComponent,
     RefreshButtonComponent,
     ExtHrefDirective,
     AuthStatusComponent,
